# storage/api-docs/openapi.yaml
openapi: 3.0.0
info:
  title: EMS Auth & Assessment API
  description: >-
    Admin & User APIs for the exam/assessment management system (Laravel 12 + Passport).
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8000/api/v1
    description: Local — mobile apps (Bearer token)
  - url: http://127.0.0.1:8000/web/v1
    description: Local — web (httpOnly cookie + CSRF)

  - url: https://seashell-okapi-169452.hostingersite.com/ems/public/web/v1
    description: Production — web (httpOnly cookie + CSRF)
  - url: https://seashell-okapi-169452.hostingersite.com/ems/public/api/v1
    description: Production — mobile apps (Bearer token)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Laravel Passport access token. Send as `Authorization: Bearer <token>`.
        Required on every endpoint except register/login/logout-adjacent public
        auth routes (marked `security: []` below). Admin tokens only authorize
        `/admin/*` routes (guard `admins`); user tokens only authorize `/user/*`
        routes (guard `users`). On the `web/v1/*` mount the same authorization can
        instead be supplied via the `admin_access_token` / `user_access_token`
        httpOnly cookie set at login — the cookie is transparently rewritten into
        this header server-side, so the documented Bearer requirement applies
        either way.
...
  schemas:
...
    AdminAssessmentListItem:
      description: >-
        One (assessment, batch) pair. Batch-wise assessments appear once per
        batch, so the same assessment can appear multiple times in a listing.
      allOf:
        - $ref: '#/components/schemas/Assessment'
        - $ref: '#/components/schemas/AssessmentBatchScheduleFull'
...
security:
  - BearerAuth: []
paths:
...
  /admin/users:
    get:
      tags: [Admin Users]
      summary: List all users
      description: >-
        Returns every user in the system (not scoped to any admin — there is no
        per-admin ownership concept for candidates). Requires admin authentication.
      operationId: adminListUsers
      responses:
        '200':
          description: List of users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Admin Users]
      summary: Create a user on a candidate's behalf
      description: >-
        Admin-created shortcut for onboarding a single candidate. Unlike
        self-registration (`/user/register`), the email is marked verified
        immediately (`email_verified_at` set to now) — no OTP step is required
        before the user can log in. Requires admin authentication.
      operationId: adminCreateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                  description: Required. Max 255 characters.
                email:
                  type: string
                  format: email
                  description: Required. Must be a valid email and unique across all users.
                password:
                  type: string
                  format: password
                  description: Required. Minimum 8 characters (matches self-registration). Stored hashed.
      responses:
        '201':
          description: User created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation failed (missing field, invalid email, or email already registered).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/batches/assessment/{assessment_id}:
    get:
      tags: [Admin Batches]
      summary: List the batches belonging to one assessment
      description: Fetch all batches for a specific assessment owned by the authenticated admin.
      operationId: adminListBatchesByAssessment
      parameters:
        - name: assessment_id
          in: path
          required: true
          description: Assessment ID, must be owned by the authenticated admin.
          schema:
            type: integer
      responses:
        '200':
          description: List of batches.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Batch'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Assessment not found or not owned by admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/batches/{id}/users:
    get:
      tags: [Admin Batches]
      summary: List users assigned to a batch
      description: >-
        Returns the users currently assigned to this batch. Like every other
        batch endpoint, the batch must belong to an assessment owned by the
        requesting admin (returns `404` otherwise) — an earlier version of
        this endpoint skipped that check, allowing any authenticated admin to
        look up any batch ID; that IDOR has been fixed. Requires admin
        authentication.
      operationId: adminListBatchUsers
      parameters:
        - name: id
          in: path
          required: true
          description: Batch ID.
          schema:
            type: integer
      responses:
        '200':
          description: Users fetched successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: The batch does not exist or its assessment is not owned by the authenticated admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Admin Batches]
      summary: Add users to a batch
      description: >-
        Assigns the given users to this batch (equivalent to calling
        `POST /admin/assignments` with this batch's `assessment_id`/`batch_id`).
        Rejected with `422` if adding them would exceed the batch's `capacity`.
        Users already assigned are not double-counted against capacity. Users
        with a scheduling conflict against another non-flexible assessment on the
        same date/time are silently skipped and reported under
        `blocked_due_to_conflict` in the response rather than causing a failure.
        Locked (`403`) once any attempt exists for this batch. Requires admin authentication.
      operationId: adminAddBatchUsers
      parameters:
        - name: id
          in: path
          required: true
          description: Batch ID, must belong to an assessment owned by the authenticated admin.
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_ids]
              properties:
                user_ids:
                  type: array
                  description: Required. IDs of existing users to assign to this batch.
                  items:
                    type: integer
      responses:
        '201':
          description: Assignment completed (some users may have been skipped — see `blocked_due_to_conflict`).
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Assignment completed
                  assigned:
                    type: array
                    items:
                      $ref: '#/components/schemas/AssessmentAssignment'
                  blocked_due_to_conflict:
                    type: array
                    description: Users skipped because they already have another non-flexible assessment overlapping this batch's schedule.
                    items:
                      type: object
                      properties:
                        user_id:
                          type: integer
                        conflicting_assessment_id:
                          type: integer
                        conflicting_assessment_title:
                          type: string
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Cannot add users after batch has started (an attempt already exists).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: The batch does not exist or does not belong to an assessment owned by the authenticated admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation failed (missing/invalid user IDs), or adding these users would exceed the batch's capacity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Admin Batches]
      summary: Remove users from a batch
      description: >-
        Removes the given users' assignments to this batch. Locked (`403`) once
        any attempt exists for this batch. Requires admin authentication.
      operationId: adminRemoveBatchUsers
      parameters:
        - name: id
          in: path
          required: true
          description: Batch ID, must belong to an assessment owned by the authenticated admin.
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_ids]
              properties:
                user_ids:
                  type: array
                  description: Required. IDs of existing users to unassign from this batch.
                  items:
                    type: integer
      responses:
        '200':
          description: Users unassigned successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Users unassigned successfully
                  deleted_count:
                    type: integer
                    description: Number of assignment rows actually removed.
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Cannot remove users after batch has started (an attempt already exists).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: The batch does not exist or does not belong to an assessment owned by the authenticated admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation failed (missing/invalid user IDs).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/assessments:
    get:
      tags: [Admin Assessments]
      summary: List all of the admin's assessments
      description: >-
        Returns one row per (assessment, batch) pair for every assessment owned
        by the authenticated admin, newest assessment first. Batch-wise
        assessments with multiple batches appear multiple times.
      operationId: adminListAssessments
      responses:
        '200':
          description: List of assessments with their batch schedules.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminAssessmentListItem'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Admin Assessments]
      summary: Create an assessment
      description: >-
        Creates a new assessment definition owned by the authenticated admin.
        Scheduling mode is fixed at creation time and cannot be changed later:
        exactly one of `is_batch_wise` / `is_flexible` may be true (both false is
        allowed and means the assessment has a single fixed-schedule batch).
        When neither flag is set, `publish_date`/`start_time`/`end_time` are
        required to seed that single batch. When `is_flexible` is true,
        `duration_minutes` is required. When `is_batch_wise` is false (including
        the flexible case), a default batch named `individual_batch_{id}` is
        auto-created immediately using the schedule fields from this request;
        batch-wise assessments start with no batches — create them separately via
        `POST /admin/batches`. Requires admin authentication.
      operationId: adminCreateAssessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - assessment_type_id
                - title
                - difficulty_level
                - is_batch_wise
              properties:
                assessment_type_id:
                  type: integer
                  description: Required. Must exist in `assessment_types`.
                  example: 1

                title:
                  type: string
                  description: Required. Max 255 characters.
                  example: Math Test

                description:
                  type: string
                  nullable: true
                  description: Optional.

                difficulty_level:
                  type: string
                  description: Required. Free text (e.g. `easy`, `medium`, `hard`) — not validated against a fixed list.
                  example: medium

                shuffle:
                  type: boolean
                  description: Optional. Randomize question order per attempt.
                  example: true

                is_library:
                  type: boolean
                  description: Optional. Marks this as a reusable template assessment.
                  example: false

                is_active:
                  type: boolean
                  description: Optional. Inactive assessments are excluded from upcoming/running/finished listings.
                  example: true

                has_negative:
                  type: boolean
                  description: Optional. Enables negative marking for MCQ auto-grading.
                  example: false

                negative_marks:
                  type: number
                  nullable: true
                  description: Optional. Minimum 0. Marks deducted per wrong MCQ answer when `has_negative` is true.
                  example: 0.5

                is_batch_wise:
                  type: boolean
                  description: Required. See `Assessment.is_batch_wise`. Cannot be true at the same time as `is_flexible`.
                  example: false

                is_flexible:
                  type: boolean
                  description: Optional (defaults false). See `Assessment.is_flexible`. Cannot be true at the same time as `is_batch_wise`.
                  example: false

                publish_date:
                  type: string
                  format: date
                  nullable: true
                  description: Required (unless `is_batch_wise` or `is_flexible` is true). Date the default batch opens.
                  example: "2026-01-30"

                start_time:
                  type: string
                  example: "10:00:00"
                  nullable: true
                  description: Required (unless `is_batch_wise` or `is_flexible` is true). Format `H:i:s`.

                end_time:
                  type: string
                  example: "11:00:00"
                  nullable: true
                  description: Required (unless `is_batch_wise` or `is_flexible` is true). Format `H:i:s`.

                expiry_date:
                  type: string
                  format: date
                  nullable: true
                  description: Optional. Only meaningful when `is_flexible` is true — last day the default batch can be attempted.
                  example: "2026-02-05"

                duration_minutes:
                  type: integer
                  nullable: true
                  description: Required when `is_flexible` is true. Minimum 1. Per-attempt time limit on the default batch.
                  example: 60

      responses:
        '201':
          description: Assessment created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Assessment created successfully
                  data:
                    $ref: '#/components/schemas/Assessment'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: >-
            Validation failed, or a scheduling business rule was violated:
            both `is_batch_wise` and `is_flexible` true, missing
            publish/start/end time for fixed-schedule assessments, or missing
            `duration_minutes` for flexible assessments.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/results/{assessment_id}/users:
    get:
      tags: [Admin Assessment Results]
      summary: List assigned users with their attempt status for an assessment
      description: >-
        One row per distinct assigned user (their most recent assignment if
        assigned more than once), flagging whether they've submitted and whether
        their attempt is still awaiting manual grading.
      operationId: adminListResultUsers
      parameters:
        - name: assessment_id
          in: path
          required: true
          description: Assessment ID, must be owned by the authenticated admin.
          schema:
            type: integer

        - $ref: '#/components/parameters/BatchIdQuery'

      responses:
        '200':
          description: List of users with attempt status.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    user_id:
                      type: integer
                    name:
                      type: string
                    attempted:
                      type: boolean
                      description: Whether this user has submitted an attempt for this assessment/batch.
                    grade_pending:
                      type: boolean
                      description: Whether the submitted attempt still has a `"Pending Evaluation"` score awaiting manual grading.
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Assessment not found or not owned by the authenticated admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: '`batch_id` is missing or invalid for a batch-wise assessment.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/results/{assessment_id}/user/{user_id}:
    get:
      tags: [Admin Assessment Results]
      summary: Get a user's final result for an assessment
      description: >-
        Full breakdown of a submitted, fully-graded attempt: score, per-question
        correctness and the choices offered. Returns `403` while any question is
        still awaiting manual grading — use the grading endpoints first.
      operationId: adminGetUserResult
      parameters:
        - name: assessment_id
          in: path
          required: true
          description: Assessment ID, must be owned by the authenticated admin.
          schema:
            type: integer

        - name: user_id
          in: path
          required: true
          description: User ID.
          schema:
            type: integer

        - $ref: '#/components/parameters/BatchIdQuery'

      responses:
        '200':
          description: User result details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  assessment_id:
                    type: integer
                  batch_id:
                    type: integer
                    nullable: true
                  user_id:
                    type: integer
                  score:
                    type: integer
                  total_marks:
                    type: integer
                  percentage:
                    type: integer
                  correct:
                    type: integer
                  wrong:
                    type: integer
                  unanswered:
                    type: integer
                  questions:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        type:
                          type: string
                          description: The assessment's type slug (all questions in one assessment share the same type).
                        question_text:
                          type: string
                        options:
                          type: array
                          description: Present (possibly empty) for every question type; only meaningful for MCQ.
                          items:
                            type: object
                            properties:
                              id:
                                type: integer
                              option:
                                type: string
                        user_answer:
                          type: string
                          nullable: true
                          description: Raw stored answer value for this question (e.g. choice text is not resolved — this is the raw `answer` JSON's scalar content where applicable).
                        is_correct:
                          type: boolean
                          nullable: true

        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: User has not submitted the assessment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Assessment results are not available yet (still pending manual evaluation).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Assessment not found or not owned by the authenticated admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: '`batch_id` is missing or invalid for a batch-wise assessment.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /user/assessments/{id}:
    get:
      tags: [User Assessments]
      summary: Get an assigned assessment
      description: >-
        Retrieve a single assessment only if it is (or was) assigned to the
        authenticated user — uses the user's most recent assignment for this
        assessment, together with its batch schedule. Requires user authentication.
      operationId: userGetAssessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Assessment ID.
      responses:
        '200':
          description: Assigned assessment details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAssessmentListItem'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Assessment not assigned to user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...